wrap network limits query in try catch, move poll for delegation logic to own hook, fix missing key
